Load submodule deps in push redirector (2nd attempt)
Summary:
A second attempt of D61582479.

D61582479 was backed out in D63384232 because it led to CPU spikes.

This was likely happening because `get_all_repo_submodule_deps` called `live_commit_sync_config.get_all_commit_sync_config_versions`, which keeps cloning a lot of configs.

This is fixed in D64104505 below by storing the repo ids in the `RepoCrossRepo` facet.

To make it worse, in the first version, the call to `get_all_repo_submodule_deps` was outside of the `if enabled` block, so it was being called **for all the repos**. Now, it will only be called for push-redirected repos.

Reviewed By: YousefSalama

Differential Revision: D64045093

fbshipit-source-id: 43665f9b95510226d1bc6410151d8d09a7f32c78