bookmarks movement: Remove unnecessary conditional
Summary:
Since we switched the strategy to run changeset hooks only on newly public changesets,
we are never going to run changeset hooks on a bookmark move that doesn't introduce new
public changesets.

The code that is being deleted here does nothing except cause unnecessary load on the phases db
and introduce an opportunity for failing a bookmark move due to a failed db read.

Reviewed By: singhsrb

Differential Revision: D64111315

fbshipit-source-id: a366a2e766aeb4b623c501029fc737534155f822