tests: add a test for pushrebase with eagerepo
Summary:
This diff adds a test for testing modern Sapling remote API push, it also demonstrates the land_stack API is failing with "not supported by the server" error.

Later, after we make land_stack API to support the test-pushrebase-eagerepo.t, we will make other push related tests to the modern push.

Context:

This diff stack will be focusing on implementing `SLAPI:land_stack` API in eagerepo, which is a blocker for using modern push in Sapling tests.

Reviewed By: quark-zju

Differential Revision: D64004790

fbshipit-source-id: 01b24d80dda7a5396e26a76d7be0304b21ac285e