Save state when operation queue is evicted
Summary:
If you have operations queued, then hit an error in the running operation, it will evict the queue, as we can't be sure your actions will make sense without the previous operation's effects.

However, we currently have no user-facing indication that this happened. We should make this explicit.

The first step is to keep track of such evicted queued operations so we can show them in the UI.

This diff keeps track of this state. This requires some idempotent storage, since we clear the queue both on error and exit events.

Reviewed By: quark-zju

Differential Revision: D64012567

fbshipit-source-id: 6ba6c8e868f93052d48df36b3cf79eb45d7825af