Recover edited commit message when queued commands are cancelled
Summary:
If you are waiting on a slow command like rebase, then do a metaedit, it will queue. If the rebase then fails, the metaedit op is not run.

If that happens, we should recover the commit message you had typed. We can put this back as a saved edited message, ready to press metaedit again.

Reviewed By: quark-zju

Differential Revision: D64017835

fbshipit-source-id: b20d37ca30e342af3bd79d77c8152f525359a5f3