initial git over slapi endpoints behind /slapigit
Summary:
We want to be able to distinguish if we're using git over edenapi or regular edenapi, but reuse all the code we can at the same time.

This puts the info which flavour of the protocol we speak in the `State` which later on handlers can read if they want to.

We can also straight away block or allow handlers.

Reviewed By: lmvasquezg

Differential Revision: D63995106

fbshipit-source-id: 95006537bd30bf71b575af39f9e90bbf3e592d40