set up extension so that WebViewInset height can be dynamically adjusted
Summary:
Context: Before we are just hard setting a number for the height for the webview inset

Use React useRef and message bus connection to extension host to dynamically send and request for extension host to adjust the inset height.

Disclosure: Current code does not work in ISL library. For it to work, we have to hard code the themeState atom from isl/src/theme.tsx to seperate the InlineComment webview from the ISL's MessageBus setup. This will be refactored at a later diff.

Reviewed By: evangrayk

Differential Revision: D63884302

fbshipit-source-id: 143c6b10a285d9b05cdb54c785b9e7866d92acfc