Using ISL SplitDiffTable to display in editor diff view
Summary:
Context: Working towards creating in editor Diff view for ISL

Summery: This is a work in progress behind a GK. Evan is still working on some refactoring for the ISL's split diff component for it to be easily reusable.

Currently for this code to work, I had to hard code the themeState atom from isl/src/theme.tsx to seperate the InlineComment webview from the ISL's MessageBus setup.

Reviewed By: evangrayk

Differential Revision: D63849533

fbshipit-source-id: 0d000793715b571198fcc6b9f8f5623bae598ba3