Make ChangedFilesWithFetching use `filePathsSubset`
Summary:
In this stack we're moving away from using `filesSubset`, which is `Array<ChangedFile>`, to just `filePathsSubset`, which is `Array<string>`.

This diff replaces the rendering of changed files in a commit in the commit info sidebar to first guess that every file path is `M` status, and then fetch the full file statuses separately. This fetch is typically pretty fast, so this does not greatly affect the user experience.

Reviewed By: quark-zju

Differential Revision: D64158492

fbshipit-source-id: 9d8b571a766802d4a9140a3c034399d91d1b5654