Make Uncommit operation take in the list of changed files with statuses
Summary: Preivously, we could get Array<ChangedFile> from the CommitInfo. Now, we only have Array<RepoRelativePath>. So we need to pass in the list of changed files to the Uncommit Operation. This is luckily quite straightforward, we already have a promise that fetches this. We can even overlap the fetch time with the confirmation interstitial.

Reviewed By: quark-zju

Differential Revision: D64158496

fbshipit-source-id: 8515e8ed2be6a290d8c0603b66d386e37e1ae4c7