Remove {file_mods} from sl log template for -85% avg `sl log` duration
Summary:
After this stack, there's no more use of path+status changed files in CommitInfo, so we no longer need this in the fetch. This is the whole reason for this migration.

Without this in the log call template, it's a significant speed up, because this info is available in the commit text and pulling and diffing the manifest is not needed.

This improves worst case and average case performance by a lot, which is very welcome. This should make ISL noticeably more reliable and snappier.

Reviewed By: quark-zju

Differential Revision: D64158495

fbshipit-source-id: 4368005e0f0903c6cb167962ef00ffd9588a08bd