Add `files` template from sl log to CommitInfo
Summary:
This stack will migrate CommitInfo's `filesSample: Array<ChangedFile>` field to be `filePathsSample: Array<RepoRelativePath>`. This allows us to use the `{files}` template in `sl log` instead of `{file_mods|json}` etc. This gives a big perf win for the `log` command because the `{files}` data is already in the commit text and doesn't require downloading and comparing manifests. For large histories, this causes timeout for some people, even after our fix to avoid fetching that data for public commits.

This diff is the starting point which adds the new fetched data without using it anywhere and without removing the old data.

Reviewed By: quark-zju

Differential Revision: D64158493

fbshipit-source-id: 8bb1839489b1d7d89ae170351531461b736a7aab