Ensure Warm Bookmarks Cache is ClientRequestInfo enabled
Summary:
Since `WarmBookmarksCache` executes in the context of the service (e.g. Mononoke Server) instead of in the context of a particular request, it is not associated with `ClientRequestInfo`. If we need to trace back problematic queries to the DB or long running blobstore operations initiated by the Warm Bookmarks Cache, we can't do that since we have no information about the origin of the WBC.

With the changes in the diff, we should now have:
- The origin service for the WBC
- Client correlator

Reviewed By: mzr

Differential Revision: D64061746

fbshipit-source-id: caf61bcb5fcfde7ba422534bdb97b4a126481fc6