teach FileAccessLogger to use StructuredLogger and add FileAccessStructuredLogger
Summary:
With this change, we no longer need `NullFileAccessLogger`, so this also gets rid of the interface and `Null` class.

Due to the removal of the interface, this needs to now live in `inodes` since the `FileAccessLogger` creates a circular dependency between `ServerState` and `InodeNumber` otherwise. I think this is okay since the logic inside of `FileAccessLogger` isn't really logging logic but rather async processing of file access requests. The logging logic is handed off to the StructuredLogger and isn't visible in this class.

Reviewed By: jdelliot

Differential Revision: D63801466

fbshipit-source-id: 1c5988f27c3df18d358aae0c35efd570daa73599