make LogEvents explictly "typed" via inheritance
Summary:
Right now "type" is strongly associated with these Event structs that we pass to the `StructuredLogger::logEvent()` method. This diff changes the structs to use inheritance to share a common expectation of a "type" member variable, rather than relying on the  `static constexpr const char* type`.

This gives us the benfit of more flexibility in the future, in particular for datasharing in the `watchman` bits, but also in preparation for "typeless" Events that will be added in a followup diff

Reviewed By: jdelliot

Differential Revision: D63797385

fbshipit-source-id: cacd8a647a4f017b89908a57caab166dd98c549c