Integration test demonstrating push failure due to missing-lfsconfig-hook
Summary:
In two AOSP repos (`aosp/manifest` and `aosp/platform/vendor/qcom/proprietary`), we saw issues with users failing to push a new branch. They would get the unexpected error message `Bookmark heads/bookmark_name does not exist` which doesn't make sense since the bookmark is being created so obviously it won't exist beforehand.

On further investigation, we realized that the problem originated from the `missing-lfsconfig-hook` which makes use of the `find_content` method. This method takes the bookmark name as input and then tries to resolve it to a changeset so that it can fetch its content. However, for a new bookmark getting created the server won't have a changeset corresponding to the bookmark which is why we get the `Bookmark does not exist` error. The integration test in this diff demonstrates this problem.

Reviewed By: mzr

Differential Revision: D64194256

fbshipit-source-id: 2997497932864b05430285dd30eb25e821c58c18