Use up deny_files hook in integration test
Summary:
To more closely reproduce production, this diff sets up a `deny_files` commit hook to block pushes to the small repo directory while the sot is still in it (i.e. the repo is not push-redirected).

Part of the SoT switch process will be to disable this commit hook.

Next diff I'll make some other small changes to get the tests even closer to what we'll do in prod.

Reviewed By: andreacampi

Differential Revision: D61597475

fbshipit-source-id: fa9dc14e2f12260bd7118982eeca10c8cea3839e