support bookmarks endpoint
Summary:
This is the first endpoint `sl clone` will hit when it's trying to clone a repo.

I'm enabling it to work for git-over-slapi. The "signature" of the request is `/bookmarks (BookmarkRequest) -> BookmarkEntry`

where

```
pub struct BookmarkRequest {
   pub bookmarks: Vec<String>,
}

pub struct BookmarkEntry {
   pub bookmark: String,
   pub hgid: Option<HgId>,
}
```

we'll reuse `HgId` as git hashes for now. We'll simply put the 20B git hash in HgId, but later on we'll change it to something like Hash20B type of structure that will hold both hg and git hashes.

Reviewed By: lmvasquezg

Differential Revision: D64063277

fbshipit-source-id: 23cd2be22a6d418b45f3e7c57df5e79e951a33f6