Support hash to location endpoint
Summary:
Enabling hash to location endpoint for git-over-slapi. The "signature" of the request is `/commit/lhash_to_location (CommitHashToLocationRequestBatch) -> CommitHashToLocationResponse`

where

```
pub struct CommitHashToLocationRequestBatch {
    pub master_heads: Vec<HgId>,
    pub hgids: Vec<HgId>,
    pub unfiltered: Option<bool>,
}

pub struct CommitHashToLocationResponse {
    pub hgid: HgId,
    pub result: Result<Option<Location<HgId>>, ServerError>,
}
```

we'll reuse `HgId` as git hashes for now. We'll simply put the 20B git hash in HgId, but later on we'll change it to something like Id20 type of structure that will hold both hg and git hashes.

Reviewed By: markbt

Differential Revision: D64134940

fbshipit-source-id: 763feed2198cc132602072e1fb47c77d5b91a8ae