Support location to hash endpoint
Summary:
Enabling location to hash endpoint for git-over-slapi. The "signature" of the request is `/commit/location_to_hash (CommitLocationToHashRequest) -> CommitLocationToHashResponse`

where

```
pub struct CommitLocationToHashRequest {
    pub location: Location<HgId>,
    pub count: u64,
}

pub struct CommitLocationToHashResponse {
    pub location: Location<HgId>,
    pub count: u64,
    pub hgids: Vec<HgId>,
}
```

we'll reuse `HgId` as git hashes for now. We'll simply put the 20B git hash in HgId, but later on we'll change it to something like Id20 type of structure that will hold both hg and git hashes.

Reviewed By: markbt, mzr

Differential Revision: D64081353

fbshipit-source-id: b0aea45939b76e7ee6639d7cc3980c6a9f3e08f7