Extract and propagate whether we recently connected to the dogfooding tier
Summary:
A single action can involve multiple server round-trips. Server affinity means we will probably end up on the same server, but that is not guaranteed. So we define "did this Response come from a special tier" as "did any of the individual requests come from a host in a special tier".
We do that by using a self-resetting boolean: setting it to `true` resets the clock; reading it after the clock has expired returns false regardless.

In practice this doesn't really matter given how we set up the config: if configured for dogfooding, Sapling will only ever talk to one host. This will hopefully change in the future to allow more flexility, but also to automatically fall back to production if dogfooding fails.
This logic will also be useful to count requests and errors talking to canary.

Reviewed By: muirdm

Differential Revision: D64061303

fbshipit-source-id: 92c645d53fd20399b2abfec27d96736e556097a7