be more consistent when accessing serialization format
Summary:
# This diff

Adds getters for the serialization format field in several structures

# Context

KeyStore contains an existing format() trait that returns the SerializationFormat. I'm planning to implement that trait for TreeStore later on, so I want to add getters for all the existing format fields so that we can access the format consistently across the codebase.

Reviewed By: quark-zju

Differential Revision: D64271012

fbshipit-source-id: f835ae736a78abd5e1dff8b731b1407f7b222fa8