Add the scaffolding for an async commit_sparse_profile_size
Summary:
## This stack

Implement an async version of `commit_sparse_profile_size` and switch the only client over.

## This diff

Implement the absolute minimum to make it compile. The `todo`s go away later in the stack.

One change of note is that the `async_requests` framework pretty much forces us to have a single params struct, so the new method accepts `CommitSparseProfileSizeParamsV2`. In theory we could make the framework more flexible, but that can come later when we have more examples.

Reviewed By: markbt

Differential Revision: D64242330

fbshipit-source-id: 628955486249c8bcc15d3e31b6f8a1a678839876