Handle redirection symlink in windows: Read redirection target mapping config in daemon
Summary:
# Background
We have seen tools like Unity messing up with already existing redirection symlink by deleting and/or creating a folder/file instead of symlink.

# Solution
When a create/delete operation is triggered for a redirected path, handle it in a way such that redirection stay intact.

# This change
Read redirection target mapping saved in checkout config as part of D63945643 and load it to CheckoutConfig.cpp so that it can be used to handle create and delete projected-fs notification on redirected path.

Differential Revision: D64151373

fbshipit-source-id: 15bec0222550d4c93da130b9593bc5bbb65f36c2