revisionstore: make trees_blocking() can be called in an async runtime
Summary:
Currently, we have two commton `block_on`s:
* 1st block_on is in python bindlings for converting Rust async to sync
* 2nd block_on is in EdenApiTreeStore.trees_blocking in Sapling Remote APIs

It is very easy to hit the nested async runtime problem when calling
Sapling Remote APIs in Rust async functions, especially when those functions
exposed to Python with `block_on`.

This diff adds `block_in_place` for the `trees_blocking()` function, so it can be
called in an async runtime.

Reviewed By: quark-zju

Differential Revision: D63612846

fbshipit-source-id: b5843118739c77117591fe32271160114621c654