Switch diagnostics to allowlist, add in formatting warnings
Summary:
Seems like the diagnostics warning is too aggressive for some LSP sources. Let's switch to an allowlist for our initial experiments, based on sources+codes that we believe to be high quality.

Additionally, we demote formatting issues to "warning" severity in the IDE, but they are typically considered land-blocking. Let's raise such warnings to be considered blocking.

We'll also adjust our logging to log both filtered error codes as well as the original unfiltered list. This data will let us refine our allowlist in the future based on signal quality. The goal should be to add in as many signals as possible to this list, or even switch it to a blocklist.

Reviewed By: madgen

Differential Revision: D64267329

fbshipit-source-id: 7205a2a2b61cb343076f391b45150ad98e7aeb1c