Integration test demonstrating issue with file mode in Mononoke Git Server
Summary:
There is a nasty bug in how Git handles file modes and since Meta has Git repositories from all over the world, we also have to handle this particular case.

Officially, Git supports only two file mode for blobs, `Regular File` with mode `100644` and `Executable File` with mode `100755`. Any user-facing Git command that you use today will eventually result in the creation of a file in either of these two modes. However, there is another file mode `100744` (executable but only for the owner) that a file can be configured with and added to some Git tree. One of those ways is by hand creating the tree (as I have done in this test) but I am sure there would be other equally creative ways as well.

Mononoke also support a file being regular or executable and it doesn't care about anything else (essentially `100744` is treated the same as `100755`). This leads to Mononoke creating Git trees that are different than the ones in Git which eventually bubbles up as clone failure.

Reviewed By: YousefSalama

Differential Revision: D64396284

fbshipit-source-id: 12cde87a794fa65ea1b87edda842689445250161