Lazily load inline comment comparison view
Summary:
If we use React.lazy, we can dynamically import this code and thus it won't be included in the main bundle, which speeds up the perf to load the inline comment webview.

Separately, we seem to have an issue with our bundler due to the inline comments webview messing with how chunks get split. I'll tackle that in future diffs, but this diff is also partially in service of that.

Reviewed By: muirdm

Differential Revision: D64406988

fbshipit-source-id: 66be5dcec650683b317bb87638776b8bfdeebcdc