move rg perf script and include it in fbpkg
Summary:
# This diff

Moves the ripgrep performance testing script to a new location and include it in the EdenFS fbpkg definition

# Context

The script was poorly placed before, as it should not be considered part of the EdenFS performance model. Instead, it's a separate script intended to be used by Virt Infra to collect performance data using their performance testing framework. We should move it to a more general location

Reviewed By: fanzeyi

Differential Revision: D64342182

fbshipit-source-id: 87b8b346691a03b74852081b3b155379f59fa6d9