remotenames: don't cache reverse rename mapping
Summary:
We cached reverse name mappings in a global variable, but this approach is incompatible with debugruntest, and in general is not good to assume there will only ever be a single configuration active for a given process.

Fix by not doing the caching. This function is called once-per-pull/push, so should be fine to not cache it.

Reviewed By: zzl0

Differential Revision: D64368266

fbshipit-source-id: e4152553089bfcdbc9d9717317837b24882b52b5