a state machine implementation for Rust Remove
Summary:
## Background
In [this doc](https://fburl.com/doc_oxidation_rm) we introduced a graph to describe how a path should be completely removed. The process is represented as a state machine.

## This Diff
In this diff, we defined some basic data structures and a very simple workflow to show how the state machine should work under this definition.

### State Impl
For each `State` variant, we define a type with the same name to hold the actual logic of what we want to do for a given state (the `next` method). And the next state will be the return value of this logic.

There are three kinds of results when "next()" runs:
1. `OK(nest_state)`: current state runs successfully; the state should tarnsition to the next state.
2. `OK(none)`: current state runs successfully; current state is one of the terminal state that there is no next state; command should exit after this state.
3. `Err`: current state sees exception and we should terminate the command and show the exception/error messages to the user.

### Main Loop
In the main loop, we keep calling `run()` method until we reach any one of the "end states" (represented by `None`).

Reviewed By: MichaelCuevas

Differential Revision: D62390953

fbshipit-source-id: fc1f5bedc584d3a2c799991499c2f882e8248898