Fix vscode extension not building in OSS
Summary:
In OSS, we don't build the inline comment webview. But it's still defined as a .html entry point. It fails when it tries to read the entry point.

We should just skip building this in OSS for now. We may later open source this, depending on if it's useful.

Reviewed By: sggutier

Differential Revision: D64431425

fbshipit-source-id: 459f4a279aa29f56213bf12460457b6d171b9a1a