tests: add a utility for executing debugruntest code in other .t tests
Summary:
We want to make it possible to run EdenFS, Mononoke, and Sapling all in the same test. In order to achieve this, we first need to make Mononoke tests somewhat compatible with debugruntest (done in D64328199). This is because EdenFS can only run in .t tests if the test supports debugruntest.

It's much easier to make it run Mononoke on debugruntest than it is to make EdenFS run outside of debugruntest, and this diff makes it possible. A good chunk of this stack will move the bare minimum functions from shell into Python for starting Mononoke under debugruntest.

Reviewed By: quark-zju

Differential Revision: D64313018

fbshipit-source-id: f7dbe41cdfbc86fe25b6c9aa4810260c68cc7e14