use hostname when remote_gitimport_lfs has twshared hostname, IP otherwise
Summary:
The tw tasks with ip-per-task enabled have ip a subjectAltName in their TLS cert. The tw tasks without ip-per-task enabled don't have subjectAltName in their TLS cert.

Because of that we have to connect using the twshred hostname to tasks without the ip-per-task and with just the IP using tasks with ip-per-task.

**There's not common way to connect.**

We want backwards compatibility for tasks with ip-per-task and tasks without it, so we can upgrade the job in TW to have ip-per-task and we don't want to `tw stop` the whole job.

Sadly the `SRChannelBuilder` always returns the parent host hostname (`twshared`) even for tasks with ip-per-task (where the hostname for a specific task on a host is more like `2.mononoke-git-server.mononoke.tsp-cln.tw.fbinfra.net`). Because of that I'm temporarily using reverse DNS resolution to get the twshared and ip-per-task hostnames so I can differentiate when I'm about to speak to a task with ip-per-task enabled or not.

We will only connect with IP once we're fully on ip-per-task.

Reviewed By: RajivTS

Differential Revision: D64416725

fbshipit-source-id: 87c6b24b640c06c948c98e0dec1578bd1b7218d0