Fix json parsing
Summary:
Fixes the issue that caused S458505 (git push failing for whatsapp android repo).

## The problem
We were using `jq` to parse the json output from SCS commands, including the command that gets the large repo that is the source of truth of a small git repo (if its SoT is not itself).

The `scsc repo-info` command with `--json` returned `push_redirected_to: null` when the SoT was in the git repo. This ran the git pushrebase code as if the SoT of the git repo was in in repo `null`, leading to scsc xrepo-lookup failures that crashed the binary.

## The fix
Use `serde_json` to properly parse the JSON.

Reviewed By: mitrandir77

Differential Revision: D64407658

fbshipit-source-id: 269331bf1e4e6bcc62200a4424c5de97d86ef80e