Enable configurable control over handling of non-standard file mode in Git
Summary:
As described in the integration test in the previous diff, we have production Git repos that use non-standard file mode in Git trees. In our current state, we realize this problem only once we have finished importing the repo and we try to clone it from Mononoke. It would be better if we had configurable control over when we want to stop the process. This diff does that.

By default, we will fail whenever we encounter non-standard file modes but the caller can override this behaviour by switching the flag in `GitImportPreferences`

Reviewed By: singhsrb

Differential Revision: D64464948

fbshipit-source-id: 9c3e7a15ddbaa2af6fcd7be9eddf11fd30e1f666