Refactor to avoid passing MononokeApp around
Summary:
## This stack

Running the worker for all repos is causing OOMs; sharding should help.

## This diff

The `MononokeApp` instance is supposed to be a singleton and remain in the top context, not passed around. Purity aside, the current code makes it hard to have multiple shards each with their executor, all holding on to a `MononokeApp` instance.

Before this diff `AsyncRequestsQueue::new` took and retained `MononokeApp` just to use it to create queues as needed; but that is actually unnecessary, we can continue using a single queue long term. This diff changes a bunch of places that used to take a "factory" to instead take a concrete queue.

Differential Revision: D64400610

fbshipit-source-id: 62b71dd07aef9b1d5e0eec9b3ce8271c64f207a7