Extract useThemeHook to SplitDiff Context
Summary:
We want to make it possible to move the split diff component to its own separate file, outside of ISL. This requires removing dependence on jotai and ISL specifics. The next import statement that causes issues here is the themeState, which transitively includes all of ISL.

By separating this, we can nearly move the diff view component out.

Reviewed By: quark-zju

Differential Revision: D64366660

fbshipit-source-id: cfa5570f93398afee9d20c60e4e13aaec09a8fc3