make comparisonInvalidatedAtom into a hook instead of an atom
Summary:
Make this API in the split diff view context into a general hook. This lets the implementor use any form of state management they want, but as a hook it is able to trigger rerenders properly.

This allows us to remove the explicit dependence on jotai in the split diff viewer, which is nice.

Reviewed By: quark-zju

Differential Revision: D64366968

fbshipit-source-id: 1ee4641684090e30682aef59ae638cee566b9051