create uri handler
Summary:
### Context
We want to add deep linking from Phabricator comment/code suggestion to the inline comment in ISL that Jack is currently working on.

### This diff
Set up the URI handler for the ISL 2.0 extension. Currently it'll take a filepath and line number and open the file. Eventually we want to also expand the comment autoamtically and also prompt user to checkout the diff if they're not already on the diff.

Reviewed By: evangrayk

Differential Revision: D64438457

fbshipit-source-id: 816f4bda8e7d4c1c5515f53de7e9285b53bbbdc9