hooks: Orchestrate running batched and non-batched hooks
Summary:
In this stack, we will batch the execution of the verify_integrity hook to take advantage
of optimization opportunities that are only available that way.

Once we do, we will have a mixture of "fast futures" (from individual file or changeset hook executions)
and "slow futures" (from batched changeset hook executions).

To run efficiently without starving the executor or running into other concurrency problems,
we join two futures: one that executes at most 10 "slow futures" at a time and one that executes
up to 100 "fast futures" at a time.

As a heuristic, it should allow us to batch 20-100 changesets at once without overwhelming the
system or causing resource starvation.

The only part of this code that is exercised for now is the non-batched hooks execution.

Reviewed By: RajivTS

Differential Revision: D64245747

fbshipit-source-id: 4cb9b249a59b37fdad9c1f2d17d6f1b2b4b4d992