hooks: Take a slice of changesets instead of an iterator
Summary:
Before this change, we inconsistently used `HashSet` or `Vec` based on where this was called from.

Instead, settle on `Vec` and slices and remove some complexity.

This stack is about enabling batching for the hooks implementation, and it makes sense to be able to use `.chunk` once we do write that implementation.

Reviewed By: RajivTS

Differential Revision: D64245746

fbshipit-source-id: 7667e1473c84c6d7c55f809979f907c3bf56b899