hooks: Propagate the concept of batched execution of changeset hooks down
Summary:
Move the logic of running multiple changesets at once all the way down so that it
can be overriden for a specific hook.

Implement the dumbest possible batched hook implementation of `verify_integrity`
(that creates one "batch" per individual changeset, running the pre-existing code)
to demonstrate how a specific hook would take advantage of the batched implementation.

The purpose of this change is to do the minimal amount of code and behavioural change
that mechanistically enables implementing a batched hook.

Later down this stack, we will implement an efficient batched implementation of the
verify_integrity hook, achieving the purpose of this entire stack.

Note: We had to move the stats logging to individual traits as that will become different once
there isn't always a one to one mapping between a changeset and a future being executed.

Reviewed By: RajivTS

Differential Revision: D64245750

fbshipit-source-id: 8653ffd54e219005f0be282fcad4f7c1fe9db6bc