open file by comment ID
Summary:
Find the filepath and repo to open with comment ID.

This way, we'll also be able to expand the target comment with the found comment handle in later diffs.

If we don't find the comment ID, we know that users most likely don't have the diff checkout at any active local repo head. In this case we can fallback to open the file with path + repo queryparam and prompt the user to checkout the diff if they want to view the comments

Reviewed By: evangrayk

Differential Revision: D64496414

fbshipit-source-id: a9e72368cbe55cc08240985dd1d19d71d5ab70e8