Move messageBus into platform
Summary:
In D64407659 I noted how the build is currently very fragile based on import ordering.

The underlying reason for this is that we have a few singleton values that we need to access in many places. This includes the message bus, clientToServerApi, and platform. The fragility came from requiring our platform to get overridden correctly when using vscode vs the browser, otherwise we'd try to use the wrong message bus or try to read from URL parameters.

We should change our structure to avoid this class of problems.

The big idea is to move all these globals into a single access location, instead of having them exist in each file. So you'd have all these globals exist inside the `platform` global.  This can be defined per platform.

This diff begins this refactor by moving messageBus into platform. This is immediately an improvement IMO. We no longer need to include both the vscode and websocket message bus implementations in both build formats. VS Code will  use the vscode webview platform, and that will include the correct message bus that uses the vscode API. The browser platform includes the websocket message bus, and not the vscode one.

Reviewed By: zzl0

Differential Revision: D64441877

fbshipit-source-id: eba07997ce1600da889adcffdd79b709b0967c43