Make vscodeWebviewPlatform replace platform.ts at bundle time
Summary:
We don't want the vscode webview to use the browser platform. Instead of using multiple entry point shenanigans, let's just replace the platform file so it has the vscode webview platform. This makes us not depend on import ordering. Callers can simply import the platform file and get the right platform.

This eliminates an entire class of issues we previously had, where changing some imports in random files could cause vscode to break entirely in production because it would load the wrong platform. Now, it's not possible to load the wrong platform.

It's still possible to put too much in the webview platform file's imports, but that's hopefully less of a big deal now.

Reviewed By: zzl0

Differential Revision: D64487275

fbshipit-source-id: 3ae0c768ac57223de266d22748a0377e1036c47f