Move urlParams into the platform
Summary:
Previously, computeUrlParams was executed as a side effect to set a global variable with the found params. This was done to clear the URL params on load, and save them to local storage / memory instead (so you don't see the token in your browser url bar).

This meant that anyone incidentally importing browersPlatformImpl would perform this url param extraction routine. In vscode, this is an error. We had a lot of weird requirements about import ordering to work around this.

Instead, by putting this computation into the BrowserPlatform, it only exists in that case. It makes a lot more sense there anyway, since these parameters are browser-specific.

Reviewed By: zzl0

Differential Revision: D64441879

fbshipit-source-id: 03dafd909656a28b31bb0629fc68bc30749a0896