Clean up the external API for new async methods
Summary:
## This stack

Miscellaneous cleanup to make the internal and external API of async methods more pleasant to use.

## This diff

Now that the first "real" method has settled, let's clean up the API for both consumers and for internal implementation.

This only touches one method. If we like it, I will also convert the `AsyncPing`. Megarepo methods can wait for a time when we can coordinate a safe rollout.

### External API

In the old API poll methods returned a struct with an optional member; its absence implicitly indicated "not ready". This diff changes the return type to a union that contains either an explicit "not ready" or an explicit success. This helps eliminate the ambiguity of what `None` might mean.

Additionally, the old poll API returned errors as a successful poll whose content pointed to an error. This has several issues:
* the error result is not logged correctly to Scuba or even accounted as an error;
* it is also invisble to ServiceRouter;
* the client needs extra steps to extract.

In the new API we unwrap any error returned by the worker and "flatten" it into a regular `ServiceError`. We also add an additional `PollError` to represent transient issues while polling. When this is returned, the client cannot make any assumption on the state of the request; retrying the poll is always safe and might return the correct result.

### Internal API

Cleans up a sore spot that made it hard to convert to the right error type in the worker.

Reviewed By: mitrandir77

Differential Revision: D64472441

fbshipit-source-id: e0b95ea21e28a287cbad68572cc79c1861427add