Paranoid check against a nonsensical setup
Summary:
## This stack

S460221 followups to fix bugs around sharding as well as general brittleness.

## This diff

In S460221 we triggered an unexpected combination (running unsharded but loading a deep-sharded config), which left us with no loaded repos. Because there was no check for that, we ended up accepting work and then failing to process it.

Let's panic instead; that would fail the Conveyor and immediately revert the release.

Reviewed By: RajivTS

Differential Revision: D64603570

fbshipit-source-id: 2c9ccf63874a813d2940ebe49576b8dbbbe4d0d3