Fix tests and apply comments from D64604922
Summary:
Some tests were already flaky, so they didn't run in D64604922.

This fixes the tests and also updates the message displayed to users when pushing to push-redirected repos.

I'll work on the branch integration test later, because D64604922 is a significant change that breaks assumptions we had in the beginning.

Reviewed By: RajivTS

Differential Revision: D64677369

fbshipit-source-id: 5f0606ce2bd4586a632ee1a8835e6315a44411eb