git_delta_manifest: derive from git trees instead of tree handles
Summary:
Switch git delta manifest derivation from depending on `TreeHandle` to using `GitTree` directly.  We can rely on this being present as we depend on `MappedGitCommitId`, which ensures that the tree that is part of the commit is present as a real tree: it was either uploaded from the user, or derived - currently via TreeHandle - when the Git commit was derived from bonsai.

The key difference here is that we are using the real raw git trees, not the ones that have been round-tripped to `TreeHandle` via bonsai, which means we have the real `mode` values for entries that have a weird mode.

Reviewed By: RajivTS

Differential Revision: D64602485

fbshipit-source-id: bc8accd6303f389d947086ad07f7a5fd7c730671