gitstore: wrap GitStore in Arc for cheap clone
Summary:
Some storemodel APIs return a `BoxIterator` that has a lifetime on `self`:

    pub type BoxIterator<'a, T> = Box<dyn Iterator<Item = T> + Send + 'a>;
    pub trait KeyStore: Send + Sync {
        fn get_..._iter(...) -> anyhow::Result<BoxIterator<anyhow::Result<(Key, Bytes)>>> { ... }
        ...
    }

The lifetime (`'a` in `BoxIterator`) is difficult to be expressed in Python
bindings. For file content fetching, to be used in Python, we cannot simply
`collect` them in a `Vec`, because of high memory usage.

In the Python bindings, a Python object for iteration purpose (`PyIter`) does
not want to have lifetime relation with something else.

Therefore, I think we should force dropping the lifetime relation on
`BoxIterator`. To do that, we need a cheap way to "copy" or "snapshot"
the store. That could be an `Arc`, starting from the `GitStore`.

Reviewed By: muirdm

Differential Revision: D64561396

fbshipit-source-id: 33fe9a921db4e0baa5aa54aa43e89f342b03b61f